Pull Request Fixes
